aoti: split Model runtime into focused TUs + PImpl/visibility, .cxx→.cpp
The single-file aoti runtime (neml2/csrc/aoti/Model.cxx, ~1575 lines) is
split into four cohesive translation units, all implementing one class:

  Model.cpp     construction (metadata parse) + the public facade
  ops.cpp       forward / jvp / jacobian
  solve.cpp     value/Newton path
  jacobian.cpp  Jacobian/IFT path

API-visibility control (header + ABI):
- Model.h is now a thin PImpl facade -- public API + an opaque
  unique_ptr<Impl>. The Segment struct, every _-helper, and all data
  members move into the non-shipped internal.h (struct Model::Impl);
  assertions.h holds the shared _throw/_assert templates.
- generate_export_header(aoti) + CXX_VISIBILITY_PRESET hidden so the
  shared library exports only the public Model API. Impl is tagged
  AOTI_NO_EXPORT (the enclosing AOTI_EXPORT class would otherwise
  propagate default visibility onto the nested type and leak its symbols).
- the generated aoti_export.h is emitted into a dedicated build-tree
  include root (never ${NEML2_BINARY_DIR} itself, which holds a stray
  `version` file that would shadow the standard <version> header).

C++ sources adopt the .cpp extension (.cxx -> .cpp). wheel.exclude drops
the raw neml2/csrc/ source tree from the wheel -- consumers read the
installed public headers from neml2/include/ and import the compiled .so;
the internals no longer ride along as dead source.

Tooling/docs updated to match: pre-commit + check_copyright (add cpp),
dependencies.yaml (Model.cxx -> Model.cpp), CMakePresets, CLAUDE.md, and
two source-comment references.

No behavior change. Verified: clean build (no warnings), export table =
public API only (0 Impl leakage), pytest tests/aoti/ 27 passed/1 skipped,
dep_manager check OK, clang-format + copyright pass, wheel inspected.
