+++ .pre-commit-config.yaml
@@ -4,7 +4,7 @@ repos:
     rev: v22.1.3
     hooks:
     -   id: clang-format
-        files: \.(cxx|h)$
+        files: \.(cxx|cpp|h)$
         pass_filenames: true
   - repo: https://github.com/astral-sh/ruff-pre-commit
     # dependencies: ruff.version
@@ -53,7 +53,7 @@ repos:
         entry: scripts/check_copyright.py --modify
         language: script
         pass_filenames: true
-        files: \.(h|cxx|py|sh|js)$
+        files: \.(h|cxx|cpp|py|sh|js)$
       - id: check-dependencies
         name: check dependency consistency
         entry: scripts/dep_manager.py check
