+++ pyproject.toml
@@ -26,6 +26,13 @@ build-dir = "build/{wheel_tag}"
 build.targets = ["aoti", "pyneml2"]
 wheel.install-dir = "neml2"
 wheel.packages = ["neml2"]
+# The C++ runtime sources under neml2/csrc/ are build inputs, not runtime
+# artifacts: the wheel imports the compiled neml2/lib/libneml2_aoti.so +
+# neml2/aoti/_aoti*.so, and C++ consumers read the *installed* public headers
+# from neml2/include/ (via the bundled pkg-config / cmake config). Excluding
+# the raw source tree keeps the PImpl internals (internal.h / assertions.h) out
+# of the shipped artifact and drops dead source weight.
+wheel.exclude = ["neml2/csrc/**"]
 
 [tool.scikit-build.cmake.define]
 NEML2_WHEEL = "ON"
